Added an event hub to handle event and callbacks
1 supports both global functions and class member functions
2 fixed some warnings and header dependence

+++ NBE/Sources/NBEWindow.cpp
@@ -7,7 +7,7 @@ namespace NBE
 		registerClass(h);
 
 		DWORD style, exStyle;
-		RECT _windowRect = { 0, 0, ri->width, ri->height };
+		RECT _windowRect = { 0, 0, LONG(ri->width), LONG(ri->height) };
 
 		if (ri->fullScreen)
 		{
